enable onnx export for INC PTQ model (#373)
* enable onnx export for PTQ

* fix output loading quantized model check

* tests refactorization

* fix style
